refactor(services): fix DID common/ imports to use correct relative paths
Update imports in files under did-manager/common/ to reference types and
schemas from the parent directory (../types.js, ../schemas.js) instead of
the incorrect sibling path (./types.js, ./schemas.js).
