+++ packages/services/src/did-manager/common/verify-did-web.test.ts
@@ -1,5 +1,5 @@
 import { verifyDidWeb } from './verify-did-web';
-import { DidVerificationCheckName } from './types';
+import { DidVerificationCheckName } from '../types';
 
 const C = DidVerificationCheckName;
